ISS27 - remove logger, add LogBook visualization
