+++ python_src/evolution/__init__.py
@@ -3,3 +3,4 @@ from .selector import Selector
 from .generator import Generator
 from .fitness import Fitness
 from .mutator import Mutator
+from .logbook import LogBook
