+++ python_src/models/__init__.py
@@ -2,5 +2,4 @@ from .car import Car
 from .script import ScriptFactory
 from .bng_vehicle import BNGVehicle
 from .road_profiler import RoadProfiler
-from .population import Population
 from .tests_generation import RoadTestFactory
